initial delete tests
