+++ .gitignore
@@ -195,3 +195,4 @@ templates/core/terraform/scripts/index.html
 templates/core/terraform/scripts/validation.txt
 templates/core/terraform/plan
 e2e_tests/pytest_e2e.xml
+e2e_tests/workspace_id.txt
